Refactor build
