+++ project/Lib.scala
@@ -1,9 +1,15 @@
 import sbt._
-import Keys._
+import sbt.Keys._
 import com.jsuereth.sbtpgp.PgpKeys._
+import org.scalajs.sbtplugin.ScalaJSPlugin.autoImport._
+import org.scalajs.jsenv.jsdomnodejs.JSDOMNodeJSEnv
 import xerial.sbt.Sonatype.autoImport._
+//import scalajsbundler.sbtplugin.ScalaJSBundlerPlugin
+//import scalajsbundler.sbtplugin.ScalaJSBundlerPlugin.autoImport._
 
 object Lib {
+  import Dependencies.Dep
+
   type PE = Project => Project
 
   val ghProject = "scalajs-react"
@@ -48,5 +54,47 @@ object Lib {
       publishTo          := Some(Resolver.file("Unused transient repository", target.value / "fakepublish")),
       packagedArtifacts  := Map.empty)
     // .disablePlugins(plugins.IvyPlugin)
+
+  def utestSettings: PE =
+    _.configure(InBrowserTesting.js)
+      .settings(
+        jsEnv                 := new JSDOMNodeJSEnv,
+        scalacOptions in Test += "-language:reflectiveCalls",
+        libraryDependencies   += Dep.MTest.value % Test,
+        testFrameworks        += new TestFramework("utest.runner.Framework"))
+
+  def extModuleName(shortName: String): PE =
+    _.settings(name := s"ext-$shortName")
+
+  def definesMacros: Project => Project =
+    _.settings(
+      scalacOptions += "-language:experimental.macros",
+      libraryDependencies ++= Seq(
+        Dep.ScalaReflect.value,
+        Dep.ScalaCompiler.value % Provided))
+
+  def hasNoTests: Project => Project =
+    _.settings(
+      fastOptJS     in Test := Attributed(artifactPath.in(fastOptJS).in(Test).value)(AttributeMap.empty),
+      fullOptJS     in Test := Attributed(artifactPath.in(fullOptJS).in(Test).value)(AttributeMap.empty),
+      sbt.Keys.test in Test := { (Test / compile).value; () },
+      testOnly      in Test := { (Test / compile).value; () },
+      testQuick     in Test := { (Test / compile).value; () })
+
+  /*
+  lazy val yarnOnPath: Boolean =
+    try {
+      Process("yarn --version").!!
+      true
+    } catch {
+      case t: Throwable => false
+    }
+
+  def useScalaJsBundler: PE =
+    _.enablePlugins(ScalaJSBundlerPlugin)
+      .settings(
+        // useYarn := yarnOnPath,
+        version in webpack := "2.6.1")
+  */
 
 }
